Moves routers to new files
